Port `run` and `KeylimeTpmError` to unique modules
This change ports the `run` function from `tpm.rs` into its own
file / module `cmd_exec.rs` - primary reason for this is that non
tpm functions also call on the `run` function, so its not
exclusive to tpm operations (for example secure_mount uses run)

For `KeylimeTpmError` - this has been ported into a module
`keylime_error`, so that we can group all errors into a single
module, as it makes sense that we will eventually have more
error types and have them neatly self contained in single module

Resolves: #81
